+++ CHANGELOG.md
@@ -2526,3 +2526,4 @@
 - [2026-07-03 12:46] refactor(security): decouple repository interface from concrete security service
 - [2026-07-03 13:13] refactor(theme): decouple repository interface from concrete theme service
 - [2026-07-03 13:13] test(state): implement widget integration tests for state
+- [2026-07-03 13:39] fix(ui): prevent double tap duplicate action in ui